consistent with dqs
