+++ src/client/api/auditLogs.ts
@@ -1,7 +1,8 @@
 import axios from "axios";
 
-import axiosConfig from "../axiosConfig";
-import { type AuditLog } from "../utils/auditLog.types";
+import { type AuditLog } from "../types/auditLog.types";
+
+import axiosConfig from "./axiosConfig";
 
 export async function getAuditLogs(): Promise<AuditLog[]> {
   const response = await axios.get("/api/audit", axiosConfig());
