+++ src/client/utils/env.ts
@@ -1,0 +1,3 @@
+export function isProduction(hostname: string): boolean {
+  return hostname.endsWith(".blaise.gcp.onsdigital.uk");
+}
