+++ src/client/utils/handleAuthRedirect.ts
